requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
black==23.12.0
click==8.1.7
Django==5.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
flake8==6.1.0
frozenlist==1.4.0
idna==3.6
inflection==0.5.1
install==1.3.5
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rest-framework-simplejwt==0.0.2
rpds-py==0.13.2
sqlparse==0.4.4
uritemplate==4.1.1
yarl==1.9.4
